Chore: 백그라운드 삭제
