--- src/pages/EventList/finishEvent.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function FinishEvent() {
-  return <div>FinishEvent</div>;
-}
